requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
python-dotenv==1.0.0
typing-extensions==4.8.0